requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cors==1.0.1
Django==5.0.6
django-cors-headers==4.3.1
django-decouple==2.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.14.0
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
psycopg2==2.9.9
pycparser==2.22
PySocks==1.7.1
python-decouple==3.8
requests==2.31.0
requests-file==2.0.0
setuptools==69.5.1
sqlparse==0.5.0
tldextract==5.1.2
tzdata==2024.1
urllib3==2.2.1
zope.event==5.0
zope.interface==6.4
